feat: update migration flow, add links
