+++ libs/prime/restake/src/components/PastClaimsCard.tsx
@@ -1,6 +1,17 @@
-import { Card, CardHeader, Stack, Typography } from '@mui/material';
+import {
+  Card,
+  CardContent,
+  CardHeader,
+  Table,
+  TableBody,
+  TableCell,
+  TableContainer,
+  TableHead,
+  TableRow,
+  Typography,
+} from '@mui/material';
 import { LrtWithdrawalStatus } from '@origin/prime/shared';
-import { ValueLabel } from '@origin/shared/components';
+import { ExternalLink, LoadingLabel } from '@origin/shared/components';
 import { tokens } from '@origin/shared/contracts';
 import { useTokenPrice } from '@origin/shared/providers';
 import { isNilOrEmpty, ZERO_ADDRESS } from '@origin/shared/utils';
@@ -12,7 +23,6 @@ import { useAccount } from 'wagmi';
 import { useUserWithdrawalsQuery } from '../queries.generated';
 
 import type { CardProps } from '@mui/material';
-import type { ValueLabelProps } from '@origin/shared/components';
 import type { Dnum } from 'dnum';
 
 export const PastClaimsCard = (props: CardProps) => {
@@ -41,66 +51,76 @@ export const PastClaimsCard = (props: CardProps) => {
       <CardHeader
         title={intl.formatMessage({ defaultMessage: 'Past Claims' })}
       />
-      <Stack pb={2}>
-        <ValueLabel
-          label={intl.formatMessage({ defaultMessage: 'Date' })}
-          value={intl.formatMessage({ defaultMessage: 'Amount' })}
-          {...valueLabelProps}
-          labelProps={{
-            variant: 'body1',
-            color: 'text.secondary',
-            ...valueLabelProps?.labelProps,
-          }}
-          valueProps={{
-            variant: 'body1',
-            color: 'text.secondary',
-            ...valueLabelProps?.labelProps,
-          }}
-        />
-        {withdrawals?.map((r) => {
-          const amt = [
-            BigInt(r?.assetAmount ?? 0),
-            tokens.mainnet.OETH.decimals,
-          ] as Dnum;
-          const converted = mul(amt, price ?? 0);
+      <CardContent>
+        <TableContainer sx={{ pb: 2 }}>
+          <Table size="small">
+            <TableHead>
+              <TableRow>
+                <TableCell>
+                  {intl.formatMessage({ defaultMessage: 'Transaction Type' })}
+                </TableCell>
+                <TableCell>
+                  {intl.formatMessage({ defaultMessage: 'Date' })}
+                </TableCell>
+                <TableCell>
+                  {intl.formatMessage({ defaultMessage: 'Amount' })}
+                </TableCell>
+              </TableRow>
+            </TableHead>
+            <TableBody>
+              {withdrawals?.map((r) => {
+                const amt = [
+                  BigInt(r?.assetAmount ?? 0),
+                  tokens.mainnet.OETH.decimals,
+                ] as Dnum;
+                const converted = mul(amt, price ?? 0);
+                const type =
+                  r.status === LrtWithdrawalStatus.Claimed
+                    ? intl.formatMessage({ defaultMessage: 'Claimed' })
+                    : intl.formatMessage({ defaultMessage: 'Migrated' });
 
-          return (
-            <ValueLabel
-              key={r.id}
-              label={intl.formatDate(new Date(r.timestamp), {
-                day: '2-digit',
-                month: '2-digit',
-                year: 'numeric',
+                return (
+                  <TableRow key={r.id}>
+                    <TableCell>{type}</TableCell>
+                    <TableCell>
+                      {intl.formatDate(new Date(r.timestamp), {
+                        day: '2-digit',
+                        month: '2-digit',
+                        year: 'numeric',
+                      })}
+                    </TableCell>
+                    <TableCell>
+                      <LoadingLabel isLoading={isPriceLoading}>
+                        {intl.formatMessage(
+                          { defaultMessage: '{amount} {converted}' },
+                          {
+                            amount: `${format(amt, 4)} OETH`,
+                            converted: (
+                              <Typography
+                                variant="body2"
+                                color="text.secondary"
+                                textAlign="start"
+                              >
+                                (${format(converted, 2)})
+                              </Typography>
+                            ),
+                          },
+                        )}
+                      </LoadingLabel>{' '}
+                    </TableCell>
+                  </TableRow>
+                );
               })}
-              value={intl.formatMessage(
-                { defaultMessage: '{amount} {converted}' },
-                {
-                  amount: `${format(amt, 4)} OETH`,
-                  converted: (
-                    <Typography
-                      variant="body2"
-                      color="text.secondary"
-                      textAlign="start"
-                    >
-                      (${format(converted, 2)})
-                    </Typography>
-                  ),
-                },
-              )}
-              isLoading={isPriceLoading}
-              {...valueLabelProps}
-            />
-          );
-        })}
-      </Stack>
+            </TableBody>
+          </Table>
+        </TableContainer>
+        <ExternalLink
+          href="https://app.yieldnest.finance/portfolio"
+          color="primary.main"
+        >
+          {intl.formatMessage({ defaultMessage: 'View on YieldNest Website' })}
+        </ExternalLink>
+      </CardContent>
     </Card>
   );
-};
-
-const valueLabelProps: Partial<ValueLabelProps> = {
-  direction: 'row',
-  px: 2,
-  pt: 1,
-  labelProps: { sx: { width: 0.5 } },
-  valueProps: { sx: { width: 0.5 } },
 };
